fix no penalty for penalty < 1

+++ src/main/java/edu/kit/kastel/sdq/artemis4j/grading/penalty/StackingPenaltyRule.java
@@ -13,7 +13,8 @@ public final class StackingPenaltyRule implements PenaltyRule {
 
     @JsonCreator
     public StackingPenaltyRule(
-            @JsonProperty(value = "penalty", required = true) int penalty, @JsonProperty("maxUses") Integer maxUses) {
+            @JsonProperty(value = "penalty", required = true) double penalty,
+            @JsonProperty("maxUses") Integer maxUses) {
         this.penalty = penalty;
         this.maxUses = maxUses;
     }
